model improvment and static files added
