+++ model/forms.py
@@ -1,19 +1,20 @@
-from django import forms
 from django.db import models
+from django.forms import ModelForm
+from model.models import SongFile
 
 
 class Instruments(models.TextChoices):
-    PIANO = "pno", "piano"
-    GUITAR = "gtr", "guitar"
-    TRUMPET = "trp", "trumpet"
+    PIANO = "piano", "piano"
+    GUITAR = "guitar", "guitar"
+    TRUMPET = "trumpet", "trumpet"
 
 
 class MusicModels(models.TextChoices):
-    MAESTRO = "mst", "maestro"
-    MOZART = "mzt", "mozart" 
-    ANIME = "anm", "anime"
-    BACH = "bch", "bach"
-    LOFI = "lof", "lofi"
+    MAESTRO = "maestro", "maestro"
+    MOZART = "mozart", "mozart" 
+    ANIME = "anime", "anime"
+    BACH = "bach", "bach"
+    LOFI = "lofi", "lofi"
 
 
 class BackgroundMusic(models.TextChoices):
@@ -22,17 +23,8 @@ class BackgroundMusic(models.TextChoices):
     THUNDER = "Thunder", "Thunder"
 
 
-class CreateSongs(forms.Form):
-    number_of_songs = forms.IntegerField()
-    song_length = forms.DecimalField()
-    background_music = forms.ChoiceField(
-        choices=BackgroundMusic.choices,
-    )
-    instrument = forms.ChoiceField(
-        choices=Instruments.choices,
-    )
-    # get a music file from the user
-    music_model = forms.ChoiceField(
-        choices=MusicModels.choices,
-    )
-
+class CreateSongs(ModelForm):
+    class Meta:
+        model = SongFile
+        exclude = ['generated_file']
+        fields = '__all__'
